trading_service/requirements.txt
fastapi==0.104.1
backtrader==1.9.78.123
pyfolio==0.9.2
empyrical==0.5.5
alpaca-trade-api==3.0.2
numpy==1.26.0
pandas==2.1.1
matplotlib==3.8.0
psycopg2-binary==2.9.9
redis==5.0.1
ccxt>=3.0.0
python-binance==1.0.19
python-dotenv==1.0.0
